AuthLambda/requirements.txt
certifi==2024.8.30
charset-normalizer==3.3.2
dataclasses==0.6
dataclasses-json==0.6.7
ecdsa==0.19.0
idna==3.8
marshmallow==3.22.0
mypy-extensions==1.0.0
packaging==24.1
pyasn1==0.6.0
python-jose==3.3.0
requests==2.32.3
rsa==4.9
six==1.16.0
typing-inspect==0.9.0
typing_extensions==4.12.2
urllib3==2.2.2
